tests: kmscapture: simple v4l2 capture test tool
Simple tool on top of libkms to display video from attached V4L cameras.
w/o parameter the kmscapture will try to use all cameras. With -s/--single
 cmd line parameter it can be forced to open only /dev/video0.
The camera resolution is chosen based on the screen resolution, number of
cameras and the supported resolutions by the camera itself.
By default the buffer is provided by DRM. This can be changed with the
--buffer-type=v4l, so the buffer is provided by the V4L layer.