utils: select() et al. need <sys/select.h>
Although indirectly included by way of other headers with glibc and
uClibc, using any of the select()-class familly of functions or macros
requires including <sys/select.h>.

This fixes the build under musl, which is way more conservative in the
ehaders that are included one-from-the-others (k.e. as few as possible).

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>