New event handling
The current event handling relies on the PageFlipHandlerBase class which
has to be implemented on the python side.

This patch implements a more versatile event handling, where any python
object can be passed as data to the commit or page flip, and it's up to
the python implementation to decide what to do with that data when
receiving the event.

Note that when doing the commit or page_flip, the ref count of the
given python object is incremented to keep it alive. The ref count is
decremented when reading the events with the new helper method
card.read_events(). This helper _has_ to be used to ensure the objects
get released properly.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>