Update to latest pybind11
Update to latest pybind11 HEAD. We can't use the latest tag (v2.2.0) as
it has a regression.

There were two problems when updating:

1) Difficulty in managing DrmObject derived classes

Most of the DrmObjects are owned by Card, and can't be allocated or
freed, but a few of them are allocated and freed by the user. For the
former, we need to use unique_ptr with py::nodelete, but that prevents
the latter from working.

The solution was to not tell the python that the latter classes derive
from DrmObject.

This seems to be missing feature in pybind11, but I think we can live
with it.

2) DrmObjects in STL containers

vector<T> where T is a DrmObject derived class doesn't work. We need to
have a manual wrapper to return vector<unique_ptr<T, py::nodelete>>
instead.

This also seems to be a pybind11 missing feature.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>