card: another try with card constructors
Let's try again with the card constructors.

Card::open_modesetting_card() is removed.

The main constructor is Card(const std::string& dev_path = "").

If dev_path is set, the device node with that path is used. If dev_path
is not set, the behavior is similar as previously, except a modeset
capable card is used at the third step:

- If KMSXX_DEVICE env variable is set, the card device with that path is
  opened.
- If KMSXX_DRIVER env variable is set, the card with the given driver
  name and index is opened. The format is either "drvname" or
  "drvname:idx".
- If neither env variable is given, the first modeset capable card is
  opened.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>