Allow making extframebuffer and dmabufframebuffer with modifiers
Many GPUs use bandwidth compression or tiling, and this information
must be passed along to KMS when constructing the framebuffer object
around the GEM handle or prime filedescriptor.

Add an vector of modifiers as an optional parameter to both of these
classes.

Bump the minimum required version of libdrm to 2.4.17 to ensure
drmModeAddFB2WithModifiers() is available.

Signed-off-by: Matt Hoosier <matt.hoosier@garmin.com>