utils: c_draw_test_pattern() takes format name, not fourcc
We can generate test patterns for pixel formats not in DRM, so we can't
take DRM fourcc as a parameter for c_draw_test_pattern(). Switch it to
format name instead.

This is a ABI change, but it is marked as super experimental.

Also, it feels a bit odd to add non-DRM format handling to kms++. But it
feels a bit pointless to split the testpat generation to a separate
library.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>